resource/waf: Only set FieldToMatch.Data if not empty
